Refactor tests
- move client to mockClient
- update service tests
- update initialize tests
- import mocked client to mocks
